fix: change signature of letsGo function - add 2-nd parameter: fragmentManager, divide hilt module file
