+++ features/home/presentation/src/main/java/com/example/presentation/ui/HomeFragment.kt
@@ -41,8 +41,6 @@ class HomeFragment : Fragment() {
             this.greeting = greeting
         }
         return inflater.inflate(R.layout.fragment_home, container, false)
-
-        //wejf;lskjf;alskdfj;l
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
@@ -57,7 +55,7 @@ class HomeFragment : Fragment() {
                 viewModel.refreshTime()
             }
             letsgoButton.setOnClickListener {
-                viewModel.letsGo(nameView.text.toString().takeIf { it != "" })
+                viewModel.letsGo(nameView.text.toString().takeIf { it != "" }, parentFragmentManager)
             }
         }
