+++ libs/core/src/main/kotlin/routing/TempRouter.kt
@@ -1,5 +1,7 @@
 package routing
 
+import androidx.fragment.app.FragmentManager
+
 interface TempRouter {
-    fun goTo_letsGo(name: String)
+    fun goTo_letsGo(name: String, manager: FragmentManager)
 }
